dev-requirements.txt
black==23.3.0
factory-boy==3.2.1
flake8==6.0.0
isort==5.12.0
mypy==1.3.0
pre-commit==3.3.3
pylint==2.17.4
pytest==7.3.1
pytest-factoryboy==2.5.1
pytest-mock==3.10.0
